release 3.9.1
This release fixes missing architectures in the XCFramework
