--- iOS_SDK/OneSignalSDK/OneSignal_XCFramework/OneSignal.xcframework/ios-i386_x86_64-simulator/OneSignal.framework/Modules/module.modulemap
@@ -1,6 +1,0 @@
-framework module OneSignal {
-  umbrella header "OneSignal.h"
-
-  export *
-  module * { export * }
-}
