--- iOS_SDK/OneSignalSDK/OneSignal_XCFramework/OneSignal.xcframework/ios-x86_64-maccatalyst/OneSignal.framework/Modules
@@ -1,1 +1,0 @@
-Versions/Current/Modules
