--- iOS_SDK/OneSignalSDK/OneSignal_XCFramework/OneSignal.xcframework/ios-x86_64-maccatalyst/OneSignal.framework/Versions/A/Modules/module.modulemap
@@ -1,6 +1,0 @@
-framework module OneSignal {
-  umbrella header "OneSignal.h"
-
-  export *
-  module * { export * }
-}
